+++ include/capability/asr_interface.hh
@@ -147,7 +147,7 @@ public:
      * @param[in] power_speech max wakeup power value
      * @param[in] callback asr recognize callback
      */
-    virtual void startRecognition(unsigned int power_noise, unsigned int power_speech, AsrRecognizeCallback callback = nullptr) = 0;
+    virtual void startRecognition(float power_noise, float power_speech, AsrRecognizeCallback callback = nullptr) = 0;
 
     /**
      * @brief Turn on the microphone and start speech recognition
